requirements.txt
albumentations==1.3.0
cityscapesScripts==2.2.1
imagecorruptions==1.1.2
imutils==0.5.4
instaboostfast==0.1.2
matplotlib==3.6.2
memory_profiler==0.61.0
mmcv==1.7.0
mmcv_full==1.7.0
numpy==1.23.5
onnxruntime==1.13.1
openai==0.25.0
opencv_python==4.6.0.66
paddleocr==2.6.1.1
pandas==1.3.5
pdf2image==1.16.0
Pillow==9.3.0
psutil==5.9.4
pycocotools==2.0.6
scikit_image==0.19.3
scikit_learn==1.1.3
scipy==1.9.3
six==1.16.0
skimage==0.0
tensorflow==2.11.0
terminaltables==3.1.10
torch==1.13.0
wandb==0.13.5
